Embettered typing and rearchitected
I removed all self-defined types and use only explicit ones, i.e. List[My_object] instead of _List_of_My_object. This is the natural way in other type systems, and is clearer to the reader.

Also I changed the architecture so that there is now a middle-man object Msg for messages. The database code creates Msg's and the API gets Msg instead of direct database ressults. That way the database  and API layers can be can be chanhed independently. It is slightly less resource efficient however since we have to create those objects as an extra step.